LPD-101584 Move the group fragment entry usages query to the service
