+++ modules/apps/fragment/fragment-web/src/main/java/com/liferay/fragment/web/internal/portlet/action/ViewGroupFragmentEntryUsagesMVCRenderCommand.java
@@ -5,8 +5,6 @@
 
 package com.liferay.fragment.web.internal.portlet.action;
 
-import com.liferay.depot.service.DepotEntryGroupRelLocalService;
-import com.liferay.depot.service.DepotEntryLocalService;
 import com.liferay.fragment.constants.FragmentPortletKeys;
 import com.liferay.fragment.service.FragmentEntryLinkLocalService;
 import com.liferay.fragment.service.FragmentEntryLocalService;
@@ -40,18 +38,11 @@ public class ViewGroupFragmentEntryUsagesMVCRenderCommand
 		renderRequest.setAttribute(
 			GroupFragmentEntryLinkDisplayContext.class.getName(),
 			new GroupFragmentEntryLinkDisplayContext(
-				_depotEntryGroupRelLocalService, _depotEntryLocalService,
 				_fragmentEntryLinkLocalService, _fragmentEntryLocalService,
 				_groupLocalService, renderRequest, renderResponse));
 
 		return "/view_group_fragment_entry_usages.jsp";
 	}
-
-	@Reference
-	private DepotEntryGroupRelLocalService _depotEntryGroupRelLocalService;
-
-	@Reference
-	private DepotEntryLocalService _depotEntryLocalService;
 
 	@Reference
 	private FragmentEntryLinkLocalService _fragmentEntryLinkLocalService;
